+++ plugins/woocommerce-admin/bin/packages/watch.js
@@ -12,7 +12,9 @@ const watch = require( 'node-watch' );
  */
 const getPackages = require( './get-packages' );
 
-const BUILD_CMD = `node ${ path.resolve( __dirname, './build.js' ).replace( /(\s+)/g, '\\$1' ) }`;
+const BUILD_CMD = `node ${ path
+	.resolve( __dirname, './build.js' )
+	.replace( /(\s+)/g, '\\$1' ) }`;
 
 let filesToBuild = new Map();
 
@@ -25,7 +27,7 @@ const exists = ( filename ) => {
 
 // Exclude deceitful source-like files, such as editor swap files.
 const isSourceFile = ( filename ) => {
-	return /.\.(js|scss)$/.test( filename );
+	return /.\.(js|ts|tsx|scss)$/.test( filename );
 };
 
 const rebuild = ( filename ) => filesToBuild.set( filename, true );
@@ -34,30 +36,48 @@ getPackages().forEach( ( p ) => {
 	const srcDir = path.resolve( p, 'src' );
 	try {
 		fs.accessSync( srcDir, fs.F_OK );
-		watch( path.resolve( p, 'src' ), { recursive: true }, ( event, filename ) => {
-			const filePath = path.resolve( srcDir, filename );
+		watch(
+			path.resolve( p, 'src' ),
+			{ recursive: true },
+			( event, filename ) => {
+				const filePath = path.resolve( srcDir, filename );
 
-			if ( ! isSourceFile( filename ) ) {
-				return;
-			}
+				if ( ! isSourceFile( filename ) ) {
+					return;
+				}
 
-			if ( ( [ 'update', 'change', 'rename' ].includes( event ) ) && exists( filePath ) ) {
-				// eslint-disable-next-line no-console
-				console.log( chalk.green( '->' ), `${ event }: ${ filename }` );
-				rebuild( filePath );
-			} else {
-				const buildFile = path.resolve( srcDir, '..', 'build', filename );
-				try {
-					fs.unlinkSync( buildFile );
-					process.stdout.write(
-						chalk.red( '  \u2022 ' ) +
-						path.relative( path.resolve( srcDir, '..', '..' ), buildFile ) +
-						' (deleted)' +
-						'\n'
+				if (
+					[ 'update', 'change', 'rename' ].includes( event ) &&
+					exists( filePath )
+				) {
+					// eslint-disable-next-line no-console
+					console.log(
+						chalk.green( '->' ),
+						`${ event }: ${ filename }`
 					);
-				} catch ( e ) {}
+					rebuild( filePath );
+				} else {
+					const buildFile = path.resolve(
+						srcDir,
+						'..',
+						'build',
+						filename
+					);
+					try {
+						fs.unlinkSync( buildFile );
+						process.stdout.write(
+							chalk.red( '  \u2022 ' ) +
+								path.relative(
+									path.resolve( srcDir, '..', '..' ),
+									buildFile
+								) +
+								' (deleted)' +
+								'\n'
+						);
+					} catch ( e ) {}
+				}
 			}
-		} );
+		);
 	} catch ( e ) {
 		// doesn't exist
 	}
@@ -68,7 +88,12 @@ setInterval( () => {
 	if ( files.length ) {
 		filesToBuild = new Map();
 		try {
-			execSync( `${ BUILD_CMD } ${ files.map( file => file.replace( /(\s+)/g, '\\$1' ) ).join( ' ' ) }`, { stdio: [ 0, 1, 2 ] } );
+			execSync(
+				`${ BUILD_CMD } ${ files
+					.map( ( file ) => file.replace( /(\s+)/g, '\\$1' ) )
+					.join( ' ' ) }`,
+				{ stdio: [ 0, 1, 2 ] }
+			);
 		} catch ( e ) {}
 	}
 }, 100 );