+++ plugins/woocommerce-admin/packages/explat/src/utils.ts
@@ -1,0 +1,4 @@
+/**
+ * Boolean determining if environment is development.
+ */
+export const isDevelopmentMode = process.env.NODE_ENV === 'development';